file_signature_recovery/requirements.txt
# ── File Signature Recovery & Analysis — Dependencies ──
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
tensorflow>=2.13.0
xgboost>=2.0.0
shap>=0.43.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
# ── Recovery System ──
yara-python>=4.3.0
psutil>=5.9.0
# ── Web API ──
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
# ── GUI ──
PyQt6>=6.5.0
